Replace some math functions with custom/different functions
The CI builds with GLIBC 2.29, but Raspberry Pi OS Buster only
supports GLIBC 2.28. To be able to run CI packages on Buster,
rewrites the math functions versioned for GLIBC 2.29.

+++ src/periphery/SFU.cpp
@@ -27,12 +27,12 @@ InstructionWalker periphery::insertSFUCall(const Register sfuReg, InstructionWal
 
 Literal periphery::precalculateSFUExp2(Literal in)
 {
-    return Literal(std::exp2(in.real()));
+    return Literal(std::exp2f(in.real()));
 }
 
 Literal periphery::precalculateSFULog2(Literal in)
 {
-    return Literal(std::log2(in.real()));
+    return Literal(std::log2f(in.real()));
 }
 
 Literal periphery::precalculateSFURecip(Literal in)
